% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prg.R
\name{plot_prg}
\alias{plot_prg}
\title{Plot the Precision-Recall-Gain curve}
\usage{
plot_prg(prg_curve, xlim = c(0, 1), ylim = c(0, 1), write_auprg = FALSE)
}
\arguments{
\item{prg_curve}{the data structure resulting from the function create_prg_curve}

\item{xlim}{the limits of the x-axis (default: c(0,1))}

\item{ylim}{the limits of the y-axis (default: c(0,1))}

\item{write_auprg}{write the area under the PRG-curve in the middle of the plot (write_auprg=TRUE) or not (write_auprg=FALSE, default)}
}
\value{
A numeric value representing the area under the Precision-Recall-Gain curve.
}
\description{
This function plots the Precision-Recall-Gain curve resulting from the function create_prg_curve using ggplot. More information on Precision-Recall-Gain curves and how to cite this work is available at http://www.cs.bris.ac.uk/~flach/PRGcurves/.
}
\details{
This function plots the Precision-Recall-Gain curve, indicating for each point whether it is a crossing-point or not (see help on create_prg_curve). By default, only the part of the curve within the unit square [0,1]x[0,1] is plotted.
}
\examples{
plot_prg(create_prg_curve(c(1,0,1,0,1,0),c(0.8,0.8,0.6,0.4,0.4,0.2)))
plot_prg(create_prg_curve(c(1,0,1,0,1,0),c(0.8,0.8,0.6,0.4,0.4,0.2)),xlim=c(-0.6,1),ylim=c(-0.6,1),write_auprg=TRUE)
}

